feat(fw): add ethereum_test_tools/utilities, clean up.
