+++ setup.cfg
@@ -67,6 +67,7 @@ lint =
     flake8>=6.1.0,<7
     pep8-naming==0.13.3
     fname8>=0.0.3
+    GitPython>=3.1
 
 docs =
     cairosvg>=2.7.0,<3  # required for social plugin (material)
