+++ src/ethereum_test_tools/spec/__init__.py
@@ -4,13 +4,7 @@ Test spec definitions and utilities.
 
 from typing import List, Type
 
-from .base.base_test import (
-    BaseFixture,
-    BaseTest,
-    TestSpec,
-    get_framework_version,
-    verify_post_alloc,
-)
+from .base.base_test import BaseFixture, BaseTest, TestSpec, verify_post_alloc
 from .blockchain.blockchain_test import BlockchainTest, BlockchainTestFiller, BlockchainTestSpec
 from .fixture_collector import FixtureCollector, TestInfo
 from .state.state_test import StateTest, StateTestFiller, StateTestOnly, StateTestSpec
@@ -31,6 +25,5 @@ __all__ = (
     "StateTestSpec",
     "TestInfo",
     "TestSpec",
-    "get_framework_version",
     "verify_post_alloc",
 )
