+++ src/ethereum_test_tools/utility/__init__.py
@@ -1,0 +1,7 @@
+"""
+Utility functions used within the EEST framework.
+"""
+
+from .helpers import get_framework_version
+
+__all__ = ("get_framework_version",)
