+++ tox.ini
@@ -10,7 +10,6 @@ development_fork = Cancun
 [testenv]
 package = wheel
 wheel_build_env = .pkg
-deps = GitPython
 
 [testenv:framework]
 description = Run checks on helper libraries and test framework
